combine into program module
