+++ pykit/__init__.py
@@ -3,5 +3,5 @@ from .ClCommand import ClCommand
 from .types import Integer
 from .types import Float
 from .types import Character
-from .Program import Program
-from .ServiceProgram import ServiceProgram
+from .program import Program
+from .program import ServiceProgram
